+++ frontend/src/pages/Main.jsx
@@ -4,6 +4,7 @@ import Sidebar from '../components/Sidebar';
 import Dashboard from '../components/Dashboard';
 import IndustryJobs from '../components/IndustryJobs';
 import CorporationJobs from '../components/CorporationJobs';
+import CorporationIndustryStats from '../components/CorporationIndustryStats';
 import Assets from '../components/Assets';
 import Planets from '../components/Planets';
 import CharacterPage from '../components/CharacterPage';
@@ -223,6 +224,8 @@ function Main({ onLogout, characterName, features = [], isAdmin = false }) {
         return <IndustryJobs onError={setError} refreshKey={refreshKey} />;
       case 'corp-jobs':
         return <CorporationJobs onError={setError} refreshKey={refreshKey} />;
+      case 'corp-industry-stats':
+        return <CorporationIndustryStats onError={setError} refreshKey={refreshKey} />;
       case 'assets':
         return <Assets onError={setError} refreshKey={refreshKey} />;
       case 'planets':